requirements.txt
# Core configuration
python-dotenv==1.0.1

# Web / API
fastapi==0.112.2
uvicorn==0.30.5

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1

# Database / Vector
psycopg2-binary==2.9.9

# Embeddings / Rerank / Models
FlagEmbedding==1.2.10
requests==2.32.3

# PDF Processing & NLP
PyMuPDF==1.24.9
camelot-py[cv]==0.11.0
pandas==2.2.2
nltk==3.9.1
tiktoken==0.7.0
pydantic==2.8.2

# Monitoring & Metrics
prometheus_client==0.19.0
Pillow==10.4.0

# Phase 2C Accuracy Improvements
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
spacy==3.6.1

# Phase 3A Multimodal Enhancement
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
ollama==0.3.1

# Optional: add version pins for frontend or other services later

# NOTE: pdf_processor/requirements.txt duplicates many of these; consider consolidating with a single root environment.
